pyproject.toml
[project]
name = "lakera-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
]
